pyproject.toml
[project]
name = "kokoro-tts"
version = "0.1.0"
description = "Text-to-Speech script using MLX Audio and Kokoro model for Apple Silicon Macs"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "google-genai>=1.67.0",
    "misaki[en]",
    "mlx-audio",
    "mutagen>=1.47.0",
    "numpy",
    "pydub",
    "python-dotenv>=1.2.2",
    "requests>=2.32.0",
    "soundfile",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-subtests",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
